requirements.txt
Package                  Version
------------------------ -----------
-umpy                    1.19.4
absl-py                  0.11.0
aiohttp                  3.6.3
astunparse               1.6.3
async-timeout            3.0.1
attrs                    20.3.0
beautifulsoup4           4.9.3
cachetools               4.1.1
certifi                  2020.11.8
cffi                     1.14.4
chardet                  3.0.4
chart-studio             1.1.0
click                    7.1.2
configparser             5.0.1
cryptography             3.3.1
cycler                   0.10.0
discord.py               1.6.0
docker-pycreds           0.4.0
Flask                    1.1.2
flatbuffers              1.12
gast                     0.3.3
gitdb                    4.0.5
GitPython                3.1.11
google-api-core          1.23.0
google-api-python-client 1.12.8
google-auth              1.23.0
google-auth-httplib2     0.0.4
google-auth-oauthlib     0.4.2
google-pasta             0.2.0
googleapis-common-protos 1.52.0
grpcio                   1.32.0
h5py                     2.10.0
httplib2                 0.18.1
idna                     2.10
importlib-metadata       3.1.1
itsdangerous             1.1.0
Jinja2                   2.11.2
joblib                   0.17.0
kaleido                  0.0.3.post1
Keras-Preprocessing      1.1.2
kiwisolver               1.3.1
Markdown                 3.3.3
MarkupSafe               1.1.1
matplotlib               3.3.3
multidict                4.7.6
nltk                     3.5
numpy                    1.19.3
oauthlib                 3.1.0
opt-einsum               3.3.0
pandas                   1.1.4
Pillow                   8.0.1
pip                      20.1.1
plotly                   4.13.0
praw                     7.1.0
prawcore                 1.5.0
promise                  2.3
protobuf                 3.14.0
psutil                   5.7.3
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pycparser                2.20
pydotplus                2.0.2
pyOpenSSL                20.0.1
pyparsing                2.4.7
python-dateutil          2.8.1
pytz                     2020.4
PyYAML                   5.3.1
regex                    2020.11.13
requests                 2.25.0
requests-oauthlib        1.3.0
retrying                 1.3.3
rsa                      4.6
sentry-sdk               0.19.5
setuptools               47.1.0
shortuuid                1.0.1
six                      1.15.0
smmap                    3.0.4
soupsieve                2.1
subprocess32             3.5.4
tensorboard              2.4.0
tensorboard-plugin-wit   1.7.0
tensorflow               2.4.0
tensorflow-estimator     2.4.0
termcolor                1.1.0
textblob                 0.15.3
tqdm                     4.54.0
typing-extensions        3.7.4.3
Unidecode                1.1.1
update-checker           0.18.0
uritemplate              3.0.1
urllib3                  1.26.2
wandb                    0.10.12
watchdog                 1.0.1
websocket-client         0.57.0
Werkzeug                 1.0.1
wheel                    0.36.1
wordcloud                1.8.1
wrapt                    1.12.1
yarl                     1.5.1
zipp                     3.4.0
